Fix #1418 and clean up
