+++ tests/func/test_start_stop.py
@@ -8,40 +8,8 @@ import platform
 import pytest
 import sys
 
-from tests.helpers.pattern import ANY, Path
+from tests.helpers.pattern import ANY
 from tests.helpers.session import DebugSession
-
-
-@pytest.mark.parametrize('start_method', ['launch'])
-def test_break_on_entry(pyfile, run_as, start_method):
-
-    @pyfile
-    def code_to_debug():
-        import backchannel
-        from dbgimporter import import_and_enable_debugger
-        import_and_enable_debugger()
-        backchannel.write_json('done')
-
-    with DebugSession() as session:
-        session.initialize(
-            target=(run_as, code_to_debug),
-            start_method=start_method,
-            debug_options=['StopOnEntry'],
-            use_backchannel=True,
-        )
-        session.start_debugging()
-
-        thread_stopped, resp_stacktrace, tid, _ = session.wait_for_thread_stopped(reason='entry')
-        frames = resp_stacktrace.body['stackFrames']
-        assert frames[0]['line'] == 1
-        assert frames[0]['source']['path'] == Path(code_to_debug)
-
-        session.send_request('continue').wait_for_response(freeze=False)
-        session.wait_for_termination()
-
-        assert session.read_json() == 'done'
-
-        session.wait_for_exit()
 
 
 @pytest.mark.parametrize('start_method', ['launch'])
