+++ tests/func/test_stop_on_entry.py
@@ -10,25 +10,32 @@ import pytest
 
 
 @pytest.mark.parametrize('start_method', ['launch'])
-def test_stop_on_entry(run_as, start_method, tmpdir):
-    testfile = tmpdir.join('test.py')
-    with testfile.open('w') as stream:
-        stream.write('''
-stop_here = 1
-print('done')
-''')
+def test_stop_on_entry(pyfile, run_as, start_method):
+
+    @pyfile
+    def code_to_debug():
+        import backchannel
+        from dbgimporter import import_and_enable_debugger
+        import_and_enable_debugger()
+        backchannel.write_json('done')
 
     with DebugSession() as session:
         session.initialize(
-            target=(run_as, str(testfile)),
+            target=(run_as, code_to_debug),
             start_method=start_method,
             debug_options=['StopOnEntry'],
+            use_backchannel=True,
         )
-
         session.start_debugging()
-        hit = session.wait_for_thread_stopped()
-        frames = hit.stacktrace.body['stackFrames']
-        assert frames[0]['source']['path'] == Path(str(testfile))
+
+        thread_stopped, resp_stacktrace, tid, _ = session.wait_for_thread_stopped(reason='entry')
+        frames = resp_stacktrace.body['stackFrames']
+        assert frames[0]['line'] == 1
+        assert frames[0]['source']['path'] == Path(code_to_debug)
 
         session.send_request('continue').wait_for_response(freeze=False)
+        session.wait_for_termination()
+
+        assert session.read_json() == 'done'
+
         session.wait_for_exit()
